refactor(docs): update import path and type definition for posts data
Change the import source of NTransform and NPost types to the correct
path in the posts.data.ts file. Also, update the type of the data
variable to align with the new type definition.

Old path: '../../src/composables/transform'
New path: '../../src/'

BREAKING CHANGE: The import paths have been updated, and the type of the
data variable has been changed. Ensure that all references to these
types and the data variable are updated accordingly.